requirements.txt
python-telegram-bot==20.0
requests==2.26.0
beautifulsoup4==4.10.0
schedule==1.1.0
